src/app/dashboard/shared/components/show-more/show-more.component.spec.ts
// import {
//   beforeEach,
//   beforeEachProviders,
//   describe,
//   expect,
//   it,
//   inject,
// } from '@angular/core/testing';
// import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
// import { Component } from '@angular/core';
// import { By } from '@angular/platform-browser';
// import { ShowMoreComponent } from './show-more.component';

// describe('Component: ShowMore', () => {
//   let builder: TestComponentBuilder;

//   beforeEachProviders(() => [ShowMoreComponent]);
//   beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
//     builder = tcb;
//   }));

//   it('should inject the component', inject([ShowMoreComponent],
//       (component: ShowMoreComponent) => {
//     expect(component).toBeTruthy();
//   }));

//   it('should create the component', inject([], () => {
//     return builder.createAsync(ShowMoreComponentTestController)
//       .then((fixture: ComponentFixture<any>) => {
//         let query = fixture.debugElement.query(By.directive(ShowMoreComponent));
//         expect(query).toBeTruthy();
//         expect(query.componentInstance).toBeTruthy();
//       });
//   }));
// });

// @Component({
//   selector: 'test',
//   template: `
//     <app-show-more></app-show-more>
//   `,
//   directives: [ShowMoreComponent]
// })
// class ShowMoreComponentTestController {
// }

